+++ gazebo/rendering/Heightmap.cc
@@ -732,7 +732,7 @@ void Heightmap::ConfigureTerrainDefaults()
     this->dataPtr->terrainGlobals->setLightMapDirection(
         Conversions::Convert(directionalLight->Direction()));
 
-    auto const &ignDiffuse = directionalLight->DiffuseColor().Ign();
+    auto const &ignDiffuse = directionalLight->DiffuseColor();
     this->dataPtr->terrainGlobals->setCompositeMapDiffuse(
         Conversions::Convert(ignDiffuse));
   }
@@ -3268,9 +3268,9 @@ Ogre::MaterialPtr TerrainMaterial::Profile::generate(
       Ogre::Vector4 splitPoints;
       const Ogre::PSSMShadowCameraSetup::SplitPointList& splitPointList =
           pssm->getSplitPoints();
-      // populate from split point 1 not 0
-      for (unsigned int t = 1u; t < numTextures; ++t)
-        splitPoints[t-1] = splitPointList[t];
+      // populate from split point 1 not 0, and include shadowFarDistance
+      for (unsigned int t = 0u; t < numTextures; ++t)
+        splitPoints[t] = splitPointList[t+1];
       params->setNamedConstant("pssmSplitPoints", splitPoints);
 
       // set up uv transform